refactor: 라우팅 경로 재정렬 및 일관화
