+++ src/routes/index.ts
@@ -4,9 +4,8 @@ import authRoutes from '#modules/auth/auth.router';
 import usersRoutes from '#modules/users/users.router';
 import projectRoutes from '#modules/projects/project.router';
 import invitationRoutes from '#modules/invitations/invitaion.router';
-import fileRoutes from '#modules/files/router';
+import fileRoutes from '#modules/files/files.router';
 import tasksRoutes from '#modules/tasks/tasks.router';
-import subtaskRoutes from '#modules/subtasks/router';
 
 const router = Router();
 
@@ -16,6 +15,5 @@ router.use('/projects', projectRoutes);
 router.use('/invitations', invitationRoutes);
 router.use('/files', fileRoutes);
 router.use('/tasks', tasksRoutes);
-router.use('/tasks/:taskId/subtasks', subtaskRoutes);
 
 export default router;
